+++ src/compas_viewer/scene/coneobject.py
@@ -1,7 +1,7 @@
 from math import pi
 from typing import Optional
-from typing import Tuple
 
+from compas.datastructures import Mesh
 from compas.geometry import Cone
 from compas.geometry import Line
 from compas.geometry import Point
@@ -27,18 +27,13 @@ class ConeObject(ViewerGeometryObject, GeometryObject):
     @property
     def points(self) -> Optional[list[Point]]:
         """The points to be shown in the viewer."""
-        pass
+        return None
 
     @property
     def lines(self) -> Optional[list[Line]]:
-        pass
+        return None
 
     @property
-    def surfaces(self) -> Optional[list[Tuple[Point, Point, Point]]]:
-        """The surface to be shown in the viewer. Currently only triangles are supported."""
-        surface_points = []
-        vertices, faces = self.geometry.to_vertices_and_faces(self.u, True)
-        for face in faces:
-            face_points = [vertices[i] for i in face]
-            surface_points.append(face_points)
-        return surface_points
+    def viewmesh(self) -> Mesh:
+        """The mesh volume to be shown in the viewer."""
+        return Mesh.from_shape(self.geometry, u=self.u, v=self.v)