+++ src/compas_viewer/scene/graphobject.py
@@ -41,6 +41,6 @@ class GraphObject(ViewerGeometryObject, BaseGraphObject):
         return [Line(self.graph.node_coordinates(u), self.graph.node_coordinates(v)) for u, v in self.graph.edges()]
 
     @property
-    def surfaces(self) -> Optional[list[Tuple[Point, Point, Point]]]:
-        """The surface to be shown in the viewer. Currently only triangles are supported."""
-        pass
+    def viewmesh(self):
+        """The mesh volume to be shown in the viewer."""
+        return None