+++ src/compas_viewer/scene/polygonobject.py
@@ -32,6 +32,6 @@ class PolygonObject(ViewerGeometryObject, GeometryObject):
         return self.geometry.lines
 
     @property
-    def surfaces(self) -> Optional[list[Tuple[Point, Point, Point]]]:
-        """The surface to be shown in the viewer. Currently only triangles are supported."""
-        pass
+    def viewmesh(self):
+        """The mesh volume to be shown in the viewer."""
+        return None